requirements.txt
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Pillow==10.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0

